Restructure demo folder and make simple layout asnd card layout wrappers
