+++ src/app/features/demo/components/demo/demo.component.html
@@ -1,24 +1,10 @@
-<div class="mx-auto max-w-3xl space-y-6 p-6">
-  <h2 class="text-2xl font-bold text-gray-800">📜 {{ 'DEMO.posts' | translate:{ count: 5 } }}</h2>
-
-  <button
-    (click)="onLoadMore()"
-    class="rounded-lg bg-blue-600 px-4 py-2 font-medium text-white shadow hover:bg-blue-700 focus:ring-2 focus:ring-blue-400 focus:outline-none"
-  >
-    {{ 'DEMO.loadNposts' | translate:{ count: 10 } }}
-  </button>
-
-  <div *ngIf="store.loading()" class="font-semibold text-blue-600">
-    {{ 'DEMO.loading' | translate }}
-  </div>
-
-  <ul class="space-y-4">
-    <li
-      *ngFor="let post of store.posts()"
-      class="rounded-lg border border-gray-200 p-4 shadow-sm transition-shadow hover:shadow-md"
+<app-simple-layout>
+  <div class="grid gap-6 lg:grid-cols-2">
+    <app-layout-card
+      title="Fetch API Demo"
+      description="Explore JSONPlaceholder posts handled by the demo store and toast feedback."
     >
-      <h3 class="mb-2 text-lg font-semibold text-gray-900">{{ post.title }}</h3>
-      <p class="text-gray-700">{{ post.body }}</p>
-    </li>
-  </ul>
-</div>
+      <app-fetch-api-demo></app-fetch-api-demo>
+    </app-layout-card>
+  </div>
+</app-simple-layout>
