+++ src/app/features/demo/components/fetch-api-demo/fetch-api-demo.component.html
@@ -1,0 +1,27 @@
+<div class="space-y-6">
+  <div class="flex flex-wrap items-center justify-between gap-4">
+    <h2 class="text-2xl font-semibold text-slate-900">
+      {{ 'DEMO.posts' | translate:{ count: 5 } }}
+    </h2>
+    <button
+      (click)="onLoadMore()"
+      class="inline-flex items-center gap-2 rounded-lg bg-blue-600 px-4 py-2 text-sm font-semibold text-white shadow-sm transition hover:bg-blue-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600"
+    >
+      {{ 'DEMO.loadNposts' | translate:{ count: 10 } }}
+    </button>
+  </div>
+
+  <div *ngIf="store.loading()" class="text-sm font-medium text-blue-600">
+    {{ 'DEMO.loading' | translate }}
+  </div>
+
+  <ul class="space-y-4">
+    <li
+      *ngFor="let post of store.posts()"
+      class="rounded-xl border border-slate-200 bg-white p-4 shadow-sm transition-shadow hover:shadow-md"
+    >
+      <h3 class="mb-2 text-lg font-semibold text-slate-900">{{ post.title }}</h3>
+      <p class="text-slate-700">{{ post.body }}</p>
+    </li>
+  </ul>
+</div>
