+++ src/app/features/demo/components/fetch-api-demo/fetch-api-demo.component.scss
@@ -1,0 +1,3 @@
+:host {
+  @apply block text-slate-900;
+}
