+++ src/app/shared/components/layout-card/layout-card.component.html
@@ -1,0 +1,15 @@
+<section class="rounded-2xl border border-slate-200 bg-white p-6 shadow-sm">
+  <div class="flex flex-col gap-6">
+    <div class="flex flex-wrap items-start justify-between gap-4" *ngIf="title || description">
+      <div class="space-y-1">
+        <h2 *ngIf="title" class="text-xl font-semibold text-slate-900">{{ title }}</h2>
+        <p *ngIf="description" class="text-sm text-slate-600">{{ description }}</p>
+      </div>
+      <ng-content select="[layout-card-actions]"></ng-content>
+    </div>
+
+    <div class="space-y-4">
+      <ng-content></ng-content>
+    </div>
+  </div>
+</section>
