+++ src/app/shared/components/layout-card/layout-card.component.scss
