+++ src/app/shared/components/layout-card/layout-card.component.ts
@@ -1,0 +1,15 @@
+import { CommonModule } from '@angular/common';
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-layout-card',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './layout-card.component.html',
+  styleUrls: ['./layout-card.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class LayoutCardComponent {
+  @Input() title = '';
+  @Input() description = '';
+}
