+++ src/app/shared/components/simple-layout/simple-layout.component.scss
