+++ src/app/shared/components/simple-layout/simple-layout.component.ts
@@ -1,0 +1,12 @@
+import { CommonModule } from '@angular/common';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'app-simple-layout',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './simple-layout.component.html',
+  styleUrls: ['./simple-layout.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class DemoAppComponent {}
